% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_Deathrate.R
\name{avg_Deathrate}
\alias{avg_Deathrate}
\title{Average Death Rate from Infectious and Parasitic Diseases for Specified Country}
\usage{
avg_Deathrate(country = NULL, data = infectious_parasitic_diseases_death_rate)
}
\arguments{
\item{country}{A character string specifying the name of the country. If NULL, calculates the average death rate for all countries.}

\item{data}{A data frame containing the death rate data, where \code{Entity} represents the country names.}
}
\value{
A numeric value representing the average death rate per 100,000 population for the specified country.
}
\description{
Average Death Rate from Infectious and Parasitic Diseases for Specified Country
}
\examples{
avg_Deathrate("Japan")
}
