% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infectious_parasitic_diseases_death_rate}
\alias{infectious_parasitic_diseases_death_rate}
\title{Age-standardized Death Rates from Infectious and Parasitic Diseases in Asian Countries}
\format{
\subsection{'infectious_parasitic_diseases_death_rate'}{

A data frame with 738 rows and 4 columns
\describe{
\item{Entity}{Country name, representing the Asian country in which the data was recorded}
\item{Code}{Country code, which is the ISO code for each country}
\item{Year}{The year in which the data was recorded}
\item{death rate per 100,000 population}{Age-standardized death rate per 100,000 population due to infectious and parasitic diseases}
}
}
}
\source{
\url{https://ourworldindata.org/grapher/infectious-and-parasitic-diseases-death-rate-who-mdb}
}
\usage{
infectious_parasitic_diseases_death_rate
}
\description{
Age-standardized Death Rates from Infectious and Parasitic Diseases in Asian Countries
}
\keyword{datasets}
